// opencdx-dashboard-new/src/api/logistics/model/manufacturer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Address } from './address';
// May contain unused imports in some cases
// @ts-ignore
import type { CategoryType } from './category-type';
// May contain unused imports in some cases
// @ts-ignore
import type { ContactInfo } from './contact-info';
// May contain unused imports in some cases
// @ts-ignore
import type { EmailAddress } from './email-address';
// May contain unused imports in some cases
// @ts-ignore
import type { PhoneNumber } from './phone-number';
// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface Manufacturer
 */
export interface Manufacturer {
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'name'?: string;
    /**
     * 
     * @type {Address}
     * @memberof Manufacturer
     */
    'manufacturerAddress'?: Address;
    /**
     * 
     * @type {ContactInfo}
     * @memberof Manufacturer
     */
    'manufacturerContact'?: ContactInfo;
    /**
     * 
     * @type {EmailAddress}
     * @memberof Manufacturer
     */
    'manufacturerEmail'?: EmailAddress;
    /**
     * 
     * @type {PhoneNumber}
     * @memberof Manufacturer
     */
    'manufacturerPhone'?: PhoneNumber;
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'manufacturerWebsite'?: string;
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'manufacturerDescription'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Manufacturer
     */
    'manufacturerCertifications'?: Array<string>;
    /**
     * 
     * @type {Timestamp}
     * @memberof Manufacturer
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof Manufacturer
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof Manufacturer
     */
    'modifier'?: string;
    /**
     * 
     * @type {Array<CategoryType>}
     * @memberof Manufacturer
     */
    'categoryType'?: Array<CategoryType>;
}